Committed Oscar's code and added rate limiting.
Adding code that Oscar wrote to visualize IMU. I modified it a little to make
the shapes work better with the card format. Somewhat of an overhaul may be
required in the near future. The good thing is that the performance is
wonderful; I upped the test client to send data at 200hz on both status and IMU,
and the renderers don't seem to make the page lag at all. In order to have the
Plotly library deal with the high rate, I added some rate limiting and now it
only refreshes once every n ms (currently 500, aka 2Hz).