Improved status box by adding log level types.
Made it so that there are log level types in the status messages. Also, the
buttons in the status bar are now generated based on the proto. Clicking on the
button will first show all messages of that log level or higher, while a second
click will filter to only those log messages. Further clicks will toggle.