Fixed an issue with the server on ws close.
In some cases, if a websocket was closed, the server would crash as it would try
to send data to a socket that wouldn't receive data. An exception was thrown but
not caught. Now, broken clients are removed. This only happened when the client
was closed but we were still trying to send data (i.e. before the onClose method
was called automatically.)