Added streaming camera support.
Camera support was added by encoding the image to JPG, converting to bytes,
adding the bytes to a proto message, and sending that message to the server. The
client then gets the bytes, converts to a base64 string, and displays the newly
created image in a canvas. This seems to work fast enough with minimal lag, but
should probably be improved as it will currently use tons of bandwidth sending a
new image every single frame. The best use case is likely to encode video on the
server with the raw images.