Adding a delta encoded camera stream.
The camera is now streamed with a delta encoding instead of simply streaming
JPGs every frame. This could be more efficient going forward, but has at the
very least pointed out a number of issues with displaying images on the canvas
due to sizing parameters. Currently, images and deltas are both encoded as PNGs
and sent as part of the protobuf message. Benchmarking is needed to figure out
exactly how much, if at all, compression is happening. It might be possible to
do some preprocessing on the images on the server to make them more uniform to
increase the amount of compression possible.