Switched system to be multiprocess.
With a little help from Ben Darnell on SO, I was able to switch the system from
being single processed to multiprocessed. In doing so, I had to switch from
using a single tornado queue to interface between the servers to using multiple
socket connections. The other servers are currently authenticating as buggies,
but that's clearly not the right approach. What likely needs to happen is a
background pubsub system needs to ferry messages between the TCP server and the
HTTP handlers. I can either build my own pubsub system, or I can use something
like Redis to do that for me. Redis has the advantage of being easy and fast,
but it's not quite what we need for a data analytics server. Writing my own will
take time but it's a learning experience.