Packaged the server deps using docker.
The docker container is currently *huge*, somewhere exceeding 2.5gb. I'll poke
around and figure out what exactly is taking up so much space, but at a guess
it's because of all of the unnecessary examples and things that are getting
built with OpenCV, as well as not deleting the protobuf source and build files.

I also don't remove any of the tools like git or cmake which are used for
building the image but are otherwise unnecessary for the actual runtime. I
imagine removing those will make the image smaller.

Also added some docker instructions to the README.