Slimmed down the Dockerfile a little.
It now includes more of OpenCV, since I'm not really sure what the final version
will end up depending on. That leaves the container with a file size of
somewhere around 1.3GB. That is quite large, but should hopefully be able to be
slimmed down later with things like clearing out the apt-cache, etc.

Interestingly, since images only ever build on top of each other, the image can
only get larger. This means that each layer has to do its own cleanup and be the
absolute minimal size that it can be. This seems like a good new person TODO.