Containerized and serving behind Nginx, sorta.
The server processes now each work inside a docker container and are served
through nginx as a reverse proxy. I still need to play with the configuration
quite a bit to get the TCP socket working in Nginx and passed to a separate
instance of the data server. Also, I need to figure out what the correct IP
address is for the http servers to talk to.

As an aside, I think this is effectively a turning point after which it'll be
inconvenient to get the system to work locally without Docker. So far, I've been
able to run everything with 0 changes so that it should work both locally and
behind Nginx, but I'm not sure that'll still be the case once I get the data
server spun up. More experimenting to come on that front.

I also added a docker-compose file which makes it super easy to spin up the
entire system (and restart as necessary). Run as follows from the docker dir:
    docker-compose up

The primary goal of all of this containerization is to make the startup process
easier for both a server and new members. It does seem like there's a bit of a
tradeoff involved in getting everything containerized, since I was previously
able to rely on Tornado's autoreload for the servers, which was super convenient
when developing. With data volumes, the files don't change as they're copied.
The source is still on my machine, but I need to do a server restart manually
with docker-compose. Oh well.