Fixed Nginx and Docker setup.
Multiple server processes are now spawned with the docker compose system, and
they all run behind Nginx as a reverse proxy. The servers then each talk to the
data server. This needs to be switched to Redis or something later, so the
servers don't need to talk to each other.

Another issue is that the test_client still runs locally rather than through a
docker container. It should be able to use the same image though, this is likely
just a configuration thing.