Fixed a few issues with the docker-compose file.
The docker-compose file was exporting too many ports (namely the ports
for the http/data_servers that didn't need to be public). Now, all of
the servers and nginx run inside their own little network, with only the
nginx ports being exposed to the localhost interface (to then be opened
to the public). Additionally, the test_client now has a few more flags
(to toggle gui and webcam), and runs in both the docker container and
locally if desired.

TODO: Figure out how to swarm this, if at all desired.

Switching the ports around in Nginx from 80:80 to something else (e.g.
8080:80) causes the websockets on the front end to stop connecting, so
I'm not entirely sure where the issue is there.