Changing symlinks to file copies.
Previously, I had symlinked all of the standard .proto files because the
webgui needed some way of getting those files. However, it's a slightly
more portable, if worse overall, solution to copy the files into this
repo directly, which is what I've done. At this point, the proto
definitions are considered stable to build off of.