Can now support multiple test clients.
I should probably change the default docker compose to spin up multiple
test clients instead of just spinning up one. Also, maybe some sort of
representation of the actual channel that data is being sent from,
though the messages don't contain any of that data so it might not be
possible. Also, right now the robot is responsible for adding its name
to ever single message, but that seems unfriendly for data usage.
Instead, I can probably send it once with the auth message, and then
persist the name for as long as that connection persists. This way the
name can be added at the server side, saving robot data. Unfortunately,
without some notion of session on the web end, it won't be possible to
save those bits ...