Build and run steps separate now.
The docker-compose file was modified so that you're now required to
build the image separately from the docker file. According to SO, a
beter design than what I was previously using is to have the image built
first, separately (e.g. via CI or on your own), and then have the
docker-comose file run the image instead of building them. This approach
is supposedly less fragile.

See link for more details:

http://stackoverflow.com/questions/40182741/build-and-use-docker-image-from-docker-compose/40185564#40185564

Additionally, I modified the auth so that here are now separate keys per
robot created via the test_client.