Logging to disk and quick converter to human fmt.
The logging needs a bunch of improvements. Currently, every single
message is being logged with a byte size as a prefix. This *works* for
the moment, but it's probably not a good idea to do thousands of disk
writes per second. Instead, some sort of batching strategy probably
makes more sense, caching messages for a while and then purging to disk
when buffers fill up.

Also, the filename probably needs some work. For now, the messages from
all robots get logged to the same file. It's probably better to have a
filtered system based on the name of the robot, and some sort of session
understanding. Right now, the system doesn't have any understanding of
session, except for whatever comes with open / close of sockets. I could
tie a session management to that, but it seems to make more sense to
consider a time based approach since it's possible that the connection
could drop. With a time based approach, the robot would have to send
some sort of heartbeat to persist its connection, maybe?

Current issue, probably won't fix: With the way docker works, with all
containers running as root, the log files created are created under
root. I could change the permissions to work for the user, but I think
it's probably unnecessary honestly as it's not that inconvient.