Added a library to communicate with the server.
Now that things are starting to stabilize a little, I've created a
little library to communicate with the server. It's kinda nice, because
it spawns off a separate thread which has its own IOLoop from tornado
running inside. This means that data transmission is completely
transparent to the end user, and that the user can even use his own
IOLoop.instance() with no issues. In the process of building this, I
learned a lot of things, including the fact that I'm an idiot ...

1. Make sure to actually satisfy the invariants, like wrapping the data
in the data queue in a Packet.
2. I need to figure out how to import files from a directory above me,
since right now the Packet file is copied (which is bad) into the lib
folder. No idea how to do that and keep it portable, though.
3. You can actually safely get a thread to commit suicide by using a
threading.Event object as a flag. By putting that in a periodic
callback, I can have the ioloop get the thread to suicide on 100ms
intervals, which is nice. Also, daemon threads are a thing.

Up next, some code cleanup and documentation.